t/t1517: indicate that `gui--askyesno` fails with GIT_TEST_INSTALLED
Commit e749c87e75 (git-gui: provide question helper for retry fallback
on Windows, 2025-08-28) causes `git-gui--askyesno` to be installed when
running `make install`. This triggers a new occurrence of the issue
noted in commit c4cf8caadd (t/t1517: mark tests that fail with
GIT_TEST_INSTALLED, 2025-08-19), which prevents Git v2.52.0 from
building in nixpkgs [1]. Indicate that this is an expected failure.

[1]: https://github.com/NixOS/nixpkgs/issues/465178#issuecomment-3580790826

Signed-off-by: Brian Lyles <brianmlyles@gmail.com>